binomotron/requirements.txt
asttokens==2.0.8
backcall==0.2.0
debugpy==1.6.3
decorator==5.1.1
entrypoints==0.4
executing==1.1.1
ipykernel==6.16.1
ipython==8.5.0
jedi==0.18.1
jupyter_client==7.4.4
jupyter_core==4.11.2
matplotlib-inline==0.1.6
mysql-connector-python==8.0.31
nest-asyncio==1.5.6
packaging==21.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.31
protobuf==3.20.1
psutil==5.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
pyzmq==24.0.1
six==1.16.0
stack-data==0.5.1
tornado==6.2
traitlets==5.5.0
wcwidth==0.2.5
